// Downloads/dpm-ui-web-and-mobile-fixes 1/dpm-ui-web-and-mobile-fixes/web-apps/consumer-portal/src/types/quoteAndBuy.ts
  export interface OwnerResponseTravel
  {
    ownerFullNameEnglish: string,
    ownerFullNameArabic: string,
    ownerDobG: string,
    ownerDobH: string,
    gender: string,
    nationality: string,
    nationalityCode: string,
    mobileNumber?: string,
    ownerId: string
  }